+++ src/components/sparklineWidget/sparklineChart.js
@@ -2,7 +2,7 @@
 import React from 'react';
 import {compose} from 'recompose';
 
-import withHighcharts from './withHighcharts';
+import withHighcharts from './../withHighcharts';
 import withSparkline from './withSparkline';
 
 import Count from './../count/count.js';
@@ -11,9 +11,7 @@ import TrendLegend from './../trendLegend/trendLegend.js';
 
 const SparklineChart = ({children, countValue, countUnits, trendLegendData}) => {
   return (
-    <div style={{border: '1px solid red', padding: '10px'}}>
-      <h1>Sparkline Chart</h1>
-
+    <div>
       {countValue && <Count value={countValue} units={countUnits} />}
       <span className="chart">{children}</span>
       {trendLegendData && trendLegendData.length > 0 && <TrendLegend data={trendLegendData} />}
