+++ src/components/sparklineWidget/sparklineWidget.js
@@ -1,23 +1,20 @@
 
 import React from 'react';
-import {compose} from 'recompose';
 
-import withHighcharts from './../withHighcharts';
-import withSparklineChart from './../withSparklineChart';
+import SparklineChart from './sparklineChart';
+
 import Tooltip from './../tooltip/tooltip';
 import {dateFormats} from './../../utils/displayFormats';
 
 import './sparklineWidget.css';
 
 
-// render a uniquely marked up and styled custom SparklineWidget
-// might also have a SparklineWidgetLarge or SparklineWidgetMonochrome
 const SparklineWidget = ({
   infoText,
   title,
   dateLastUpdated,
-  children,
   viewport,
+  series, xAxis
 }) => {
   return (
     <article role="article" className="D_widget">
@@ -26,15 +23,11 @@ const SparklineWidget = ({
         <h1 className="highcharts-title">{title}</h1>
         <span className="highcharts-subtitle">Last updated at <time dateTime={dateFormats.dateTime(dateLastUpdated)}>{dateFormats.dayMonthYear(dateLastUpdated)}</time></span>
       </header>
-      <section>{children}</section>
+      <section>
+        <SparklineChart series={series} xAxis={xAxis} />
+      </section>
     </article>
   )
 };
 
-const enhance = compose(
-  withHighcharts,
-  withSparklineChart
-)(SparklineWidget);
-
-export default enhance;
-
+export default SparklineWidget;
